Move defaultfactory to src/default
